Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This works on a PPC iBook G4, a Leopard 32-bit VM, and a Snow Leopard MBP.

The binary sizes between the other and new compilation are quite different, so I'm not sure it was compiled with the same configuration. I'm not sure it matters that much, though.

All the other executables are i386-only, so I'm not sure what the policy is going forward.

Patch-set: 1
Reviewer: Kenny Root <1000748@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1