Add OverrideCheck doclet to prebuilt to support builds on javac 1.6
This adds a doclet called OverrideCheck which finds interface method @Override annotations which cause a build with javac 1.5 to crash,  but aren't caught by javac 1.6 when using -source 1.5. This javac inconsistency appears to be the main blocker to allowing java 1.6 platform build environments. A 2nd patch for build/core makes use of this tool to catch these annotations.

OS X 10.6 (Snow Leopard) no longer includes java 1.5, so overcoming this hurdle is key to ease of platform development in that environment.

More info on the doclet, written by Tom Gibara, is available at:
http://www.tomgibara.com/android/override-check

Relevent discussion links:
http://groups.google.com/group/android-platform/browse_thread/thread/bdedb17dbc349996/

http://groups.google.com/group/android-platform/browse_thread/thread/8b565af4a0ef82d8