Update patch set 1
Patch Set 1: No score

To make sure I'm clear on the request:

I'll submit a separate patch to build/tools, moving the source to that patch set and adding explicit build instructions.

This patch will be updated to contain only the binary (and text files), excluding the src/ folder and referencing the build/tools as an aditional option for retrieving the source.

Tom, given the nature of the annotation that the doclet is watching for, do you happen to know if it will still work properly if compiled with java 1.5?  If not, I'll still proceed with the requested changes, but call out the fact that it must be compiled with java 1.6 in the build notes.

Patch-set: 1