Update patch set 1
Patch Set 1: No score

If the doclet will compile with both versions of java, I'd really prefer that it not be checked in prebuilt, but instead have a dependency added so that it's just built as part of the build.  Prebuilts are a pain to deal with for many reasons, not the least of which is that people forget how to build them.

Patch-set: 1
Reviewer: Joe Onorato <1000701@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0