Update patch set 1
Thanks for all the feedback.

I'll look into what it would take to update patch 12074 to avoid using a prebuilt binary as soon as I can.

If I hit a roadblock or it looks like it will really slow progress down, I'll proceed with JBQ's initial feedback to keep things moving forward with a prebuilt/build combo and revisit the issue again to avoid the prebuilt binary at a later date if possible.

Patch-set: 1