i686-linux-glibc2.7-4.4.3: Fix rebuild script.
This patch fixes the rebuild script to not use hard-coded Ubuntu
package URLs. This was problematic because some of these became
invalid when Ubuntu increased the version number of some of these
packages, and moved various ones to security.ubuntu.com.

Instead, we know download the Packages.bz2 corresponding to a given
release and parse it to extract the versioned .deb URL to use from
the mirror.

Also add command-line options parsing. You don't need to define NDK
in your environment anymore. Instead, pass the path to the toolchain
sources directory, as in:

   <progname> /path/to/toolchain/src

Alternatively, use --ndk-dir=<path> to specify the path to your NDK,
and the script will automatically download and patch the toolchain
sources from android.git.kernel.org for you, as in:

   <progname> --ndk-dir=/path/to/ndk

By default, the script creates a tarball now (/tmp/<archive-name>),
but you can use --prefix=<path> to indicate an installation directory,
as in:

   <progname> --prefix=/opt/workspace/prebuilt/linux-x86/toolchain/ <other-options>

IMPORTANT NOTE: The generated toolchain binaries still do not run
properly on Glibc 2.7, only the machine code they generate. We hope
to solve this in the future with another patch.

Use --help for complete usage information.

Change-Id: I89390e0121438440bc756eac66ed108f0e6697a8