Update patch set 1
Patch Set 1: Fails

Actually, there are going to be issues with the makefile: this causes the files in question to be installed in all builds, and I'm especially worried that we're planning to use GPL files there instead.

That's why in fact ALL_PREBUILT is not allowed in honeycomb, to prevent such situations.

Ying, what's the correct approach to copy files around? I'd suggest PRODUCT_COPY_FILES in the device-specific side of the product definition, but there might be other options.

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1