Update patch set 1
Patch Set 1:

Do they even need to be copied anywhere?

There's logic in the emulator (when running in the dev tree) for it to go look for system images in specific places. can't we do the same for the emulator to go look for the bios files in prebuilt/android-x86/... ?

Then for the SDK build we can copy them from prebuilt into the SDK folder only.

Patch-set: 1
Label: Verified=0