Update patch set 1
Patch Set 1:

The idea is that external/qemu-pc-bios is hand built when required and the binarys are copied to this prebuilt directory and committed here. The SDK build process will install the binaries from this directory.

I've received some suggestions on how to re-write this Android.mk and will give it a go.

Patch-set: 1