emulator kernel: Fix SurfaceFlinger flickering.
The kernel images provided in this patch were built by using the sources
from android.git.kernel.org/kernel/qemu.git with the following hash:

    b0d93fb0426911d0329f861f22c59f1c72cff815

And changing the files goldfish_defconfig and goldfish_armv7_defconfig
under arch/arm/configs/ in the following way:

- Disable CONFIG_FRAMEBUFFER_CONSOLE
- Enable FB_EARLYSUSPEND

This is required to get rid of an annoying flickering problem in SurfaceFlinger
when running the system within the emulator.

See bug 4120365 for details.

Change-Id: I85d19c2a26d7fd45aebe8754c20016368d6c1422