i686-linux-glibc2.7-4.4.3: Add missing sysroot libraries.
This adds a few missing libraries to our custom 32-bit toolchain sysroot in order
to be able to build the OpenGL Hardware emulation test program (and later complete
implementation).

You can check this by defining BUILD_EMULATOR_OPENGL=true in your environment.
After this patch is submitted, we will be able to remove the hack under the
following build script:

  developement/tools/emulator/opengl/tests/ut_renderer/Android.mk

+ Add a --only-sysroot option to the build script that can be used to
  only rebuild a new sysroot, without recompiling the full toolchain,
  which can be quite long.

Change-Id: Ic6a41e90932c3595245b58701e06e5fb7d7a3519