goldfish: x86: Fix framebuffer power management.
This patch adds new prebuilt kernel images that support proper
power management of the virtual framebuffer device. This is needed
because the x86 defconfig was missing a few critical definitions
(no such patch is needed for the ARM kernel images).

These images were built from the sources on android.git.kernel.org/kernel/qemu.git
corresponding to the following hash:

    95347fcbe58b521d08b132e0d109beed7d4f11b5

In addition, the following settings were manually set in the corresponding defconfig
files to support HC's SurfaceFlinger changes (which are not in AOSP / GB):

   CONFIG_FB_EARLYSUSPEND=y
  # CONFIG_FRAMEBUFFER_CONSOLE is not set

Change-Id: I672acc84854c9efe912895eb4e780ae3e8f97f7d