sdk: new prebuilt Linux SDK toolchain
This patch updates the prebuilt binaries of the special toolchain
used by the linux sdk build (and associated sysroot). They are
used to ensure that the generated host binaries run on Ubuntu 8.04
and other systems with an old version of the GNU C Library.

The build-hardy-toolchain.sh script (provided here) has
been considerably improved:

- it doesn't require an external NDK installation anymore. It will
  download the toolchain sources from the right git repositories
  on its own. See the new --git-base and --git-reference options
  in the --help.

- it can generate a bootstrapped toolchain with the --bootstrap
  option, i.e. one that is compiled with itself. The benefit is
  that the resulting binaries also run on 8.04.

- it creates two files (PACKAGE_SOURCES and TOOLCHAIN_SOURCES)
  that given the SHA1 numbers of all downloaded Ubuntu packages
  and of the toolchain git source repositories.

- it copies itself into the generated archive.

IMPORTANT NOTE: This is _not_ a bootstrapped toolchain because
using it breaks the SDK build.

The reason for this is that we use some host prebuilts (e.g.
libbfd.a) which have been generated with another toolchain.
The library then fails to link, when the linker complains that
it can't find __asprintf_chk, which apparently is a new
addition to GLibc, not provided by our sysroot.

Fixing this would require rebuilding these prebuilts with
this toolchain first. This is left as an exercise to the
reader.

Change-Id: Idd0926af258aef77a88df1b054da60699d746bfa